back-dine/dine_api/requirements.txt
asgiref==3.3.1
cached-property==1.5.2
click==7.1.2
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.5.0
django-filter==2.1.0
django-url-filter==0.3.15
django-utils-six==2.0
djangorestframework==3.9.3
drf-extensions==0.6.0
enum-compat==0.0.3
gunicorn==20.0.4
Markdown==3.3.3
orderdict==2020.12.3
psycopg2==2.8.6
pytz==2020.4
setupcfg==2020.12.3
six==1.15.0
sqlparse==0.4.1
values==2020.12.3
whitenoise==5.2.0
